Removed old bookstore-se app and renamed bookstore-nima to bookstore-se.
