--- tests/apps/bookstore/bookstore-nima/src/main/resources/application.yaml
@@ -1,30 +1,0 @@
-#
-# Copyright (c) 2019, 2023 Oracle and/or its affiliates.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-app:
-  json-library: jsonp
-
-server:
-  port: 8080
-  host: 0.0.0.0
-#  ssl:
-#    private-key:
-#      keystore-resource-path: "certificate.p12"
-#      keystore-passphrase: "helidon"
-#  experimental:
-#    http2:
-#      enable: true
-#      max-content-length: 16384
