+++ tests/apps/bookstore/bookstore-se/src/main/resources/application.yaml
@@ -1,5 +1,5 @@
 #
-# Copyright (c) 2019, 2021 Oracle and/or its affiliates.
+# Copyright (c) 2019, 2023 Oracle and/or its affiliates.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -24,7 +24,4 @@ server:
 #    private-key:
 #      keystore-resource-path: "certificate.p12"
 #      keystore-passphrase: "helidon"
-#  experimental:
-#    http2:
-#      enable: true
-#      max-content-length: 16384
+
