+++ tests/apps/bookstore/bookstore-se/src/test/java/io/helidon/tests/apps/bookstore/se/Http2SslTest.java
@@ -23,8 +23,8 @@ import java.util.concurrent.ExecutorService;
 import java.util.concurrent.TimeUnit;
 
 import io.helidon.common.configurable.ThreadPoolSupplier;
-import io.helidon.reactive.webserver.WebServer;
 
+import io.helidon.nima.webserver.WebServer;
 import okhttp3.OkHttpClient;
 import okhttp3.Protocol;
 import okhttp3.Request;
