+++ tests/apps/bookstore/pom.xml
@@ -35,7 +35,6 @@
 
     <modules>
         <module>common</module>
-        <module>bookstore-nima</module>
         <module>bookstore-se</module>
         <module>bookstore-mp</module>
     </modules>
